Add SnapshotController as standalone, toggleable feature with full state handling
